feat(audio): Improve audio handling and enforce SoundId constraints
- Added Mix_Init(0) to support future audio formats and wrapped loading in try-catch for resource management.
- Enforced compile-time checks for SoundId ordering to prevent silent errors.
- Fixed audio playback logic to ensure proper resource cleanup and prevent leaks.
- Adjusted ghost behavior and collision detection to enhance gameplay experience.
- Updated rendering logic to use actual movement direction for smoother visuals.
- Corrected various bugs related to ghost modes and player interactions.

+++ src/util/grid.hpp
@@ -36,7 +36,7 @@ public:
   }
 
   bool outOfRange(const std::size_t i) const {
-    return i >= area();
+    return i >= static_cast<std::size_t>(area());
   }
 
   bool outOfRange(const Pos pos) const {
